+++ internal/types/upgrade_info.go
@@ -1,0 +1,8 @@
+package types
+
+// UpgradeInfo is the details from the regexp
+type UpgradeInfo struct {
+	Name    string
+	Height  int
+	Version string
+}
