--- runner/testdata/validate/runner/genesis/bin/test-runnerd
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-echo Genesis $@
-sleep 1
-echo 'binary is running, upgrade pending'
-sleep 2
-echo Never should be printed!!!
