+++ x/runner/common_test.go
@@ -1,0 +1,33 @@
+package runner
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+
+	"github.com/pkg/errors"
+	"github.com/pokt-network/pocket-runner/internal/types"
+)
+
+// copyTestData will make a tempdir and then
+// "cp -r" a subdirectory under testdata there
+// returns the directory (which can now be used as config.Config.Home) and modified safely
+func copyTestData(subdir string) (string, error) {
+	tmpdir, err := ioutil.TempDir("", "pocket-runner-test")
+	if err != nil {
+		return "", errors.Wrap(err, "create temp dir")
+	}
+
+	src := filepath.Join("testdata", subdir)
+
+	options := types.Options{
+		Recursive: true,
+		Atomic:    true,
+	}
+	err = types.Copy(src, tmpdir, options)
+	if err != nil {
+		os.RemoveAll(tmpdir)
+		return "", errors.Wrap(err, "copying files")
+	}
+	return tmpdir, nil
+}
